First stab at browser-based auth
This is an attempt at copying the Github CLI's awesome web-based CLI
authentication, which I thought was super clever and intuitive.  This
doesn't work yet, as it turns out the Linode OAuth service returns the
token in the URL hash, which isn't sent to the server.  I'd rather not
use the private OAuth flow, as the secret would need to be embedded in
the CLI and clearly wouldn't be secret.  I'm not sure it can work, but
I'm going to think on it.